backend/app/requirements.txt
fastapi
uvicorn[standard]
boto3
passlib[bcrypt]
pyjwt
python-dotenv